pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "GenAI Unit Test"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "certifi>=2025.8.3",
    "deadcode>=2.4.1",
    "dotenv>=0.9.9",
    "langgraph>=0.6.7",
    "litestar>=2.17.0",
    "ollama>=0.6.0",
    "openai>=1.109.1",
    "psycopg2-binary>=2.9.11",
    "qdrant-client>=1.15.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "structlog>=25.4.0",
    "transformers==4.57.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
    "typing-extensions>=4.15.0",
    "uvicorn>=0.37.0",
]
